// web-app/src/app/types/employee.ts
export type Employee = {
  id: number;
  firstName: string;
  lastName: string;
  email: string;
  department: string;
  jobTitle: string;
  salary: number;
  hireDate: string;
  age: number;
  city: string;
  country: string;
  isActive: boolean;
  createdAt: string;
  updatedAt: string;
  projects: unknown[];
};